Update TagDao to work with gift_certificate_has_tag table
