+++ GiftCertificatesSystem/persistence/src/main/java/com/epam/esm/dao/GiftCertificateDao.java
@@ -1,12 +1,6 @@
 package com.epam.esm.dao;
 
 import com.epam.esm.entity.GiftCertificate;
-import com.epam.esm.entity.Tag;
-
-import java.util.List;
 
 public interface GiftCertificateDao extends CrudDao<GiftCertificate> {
-    void saveGiftCertificateIdAndTagId(Long giftCertificateId, Long tagId);
-
-    List<Tag> findAllTagsByGiftCertificateId(Long id);
 }
