+++ GiftCertificatesSystem/persistence/src/main/java/com/epam/esm/dao/mapper/TagMapper.java
@@ -9,11 +9,10 @@ import java.sql.SQLException;
 public class TagMapper implements RowMapper<Tag> {
     @Override
     public Tag mapRow(ResultSet rs, int rowNum) throws SQLException {
-        Tag tag = new Tag();
-        tag.setId(rs.getLong("id"));
-        tag.setName(rs.getString("name"));
+        Tag tag = Tag.builder()
+                .id(rs.getLong("id"))
+                .name(rs.getString("name"))
+                .build();
         return tag;
     }
-
-
 }
