+++ GiftCertificatesSystem/web/src/main/java/com/epam/esm/context/ApplicationConfig.java
@@ -1,9 +1,5 @@
 package com.epam.esm.context;
 
-import com.epam.esm.dao.GiftCertificateDao;
-import com.epam.esm.dao.TagDao;
-import com.epam.esm.dao.impl.GiftCertificateDaoImpl;
-import com.epam.esm.dao.impl.TagDaoImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
